fixed logger & move transcribe_voice_message func to utils
