+++ app.py
@@ -1,2 +1,4 @@
+from logs.logging import logger
+logger.info('The app is running...')
 from bot import init_telegram_bot # run telegram bot server
