+++ bot/utils.py
@@ -4,6 +4,28 @@ from pydub import AudioSegment
 import sys
 sys.path.append("../")
 from logs.logging import logger
+from os import path
+
+def transcribe_voice_message(file_id):
+    try:
+        # Load the OGG file
+        ogg_audio = AudioSegment.from_file(f"bot/voice_messages/{file_id}.ogg", format="ogg")
+
+        # Export it as WAV
+        ogg_audio.export(f"bot/voice_messages/{file_id}.wav", format="wav")
+
+        # Transcribe audio file
+        AUDIO_FILE = f"bot/voice_messages/{file_id}.wav"
+
+        # Use the audio file as the audio source
+        r = sr.Recognizer()
+        with sr.AudioFile(AUDIO_FILE) as source:
+            audio = r.record(source)  # Read the entire audio file
+
+        transcription = r.recognize_google(audio)
+        logger.info(f"Transcription for file {file_id}: {transcription}")
+    except Exception as e:
+        logger.error(f"Error transcribing file {file_id}: {str(e)}")
 
 def convert_text_to_audio(text: str, output_file: str) -> None:
     try:
@@ -51,21 +73,3 @@ def convert_mp3_to_wav(mp3_file: str, wav_file: str) -> None:
         logger.info("MP3 file converted to WAV: %s", wav_file)
     except Exception as e:
         logger.error("An error occurred while converting MP3 to WAV: %s", str(e))
-
-if __name__ == "__main__":
-    text_to_convert = "Hey, my name is Lior Atyia, and I am your English tutor"
-    mp3_output_file = "output.mp3"
-    wav_output_file = "output.wav"
-
-    # Convert text to audio
-    convert_text_to_audio(text_to_convert, mp3_output_file)
-
-    # Convert MP3 to WAV
-    convert_mp3_to_wav(mp3_output_file, wav_output_file)
-
-    # Extract text from the WAV file
-    result = convert_audio_to_text(wav_output_file)
-
-    if result:
-        print("Converted Text:")
-        print(result)
